refkit image: whole-disk encryption with TPM NVRAM key
The per-machine key is stored in TPM NVRAM, where it is only
available to valid, authorized software once Secure Boot is
enabled.

This is a proof-of-concept. It relies heavily on working Secure Boot,
which requires further work and solving some usability questions, like
distinguishing between a device fresh from the factory and a locked
down device.

Bringing up TPM inside the initramfs depends on IPv4. init-ifupdown is
used to provide the necessary config files.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>