refkit image: support dm-verity
dm-verity is useful as an integrity protection mechanism for read-only
partitions. In contrast to whole-disk encryption, it also works for
removable media like USB sticks. Accidental (media errors) or
intentional (offline attacks) modifications are detected at runtime
directly when accessing blocks. This would be useful for the installer
image (more reliable flashing), even when not worrying about offline
attacks.

However, the distro feature is currently disabled by default because it
depends on pending patches in meta-oe which enable building the
necessary tools also natively.

A default signing key is provided as part of this change, but only
used when the build is configured to be for development. When building
a production image, the developer needs to create and configure his
own secret key.

OpenSSL is used for signinging and verification. GnuPG was considered.
It would make passphrase handling for the private key a bit more
flexible, but adds a huge chain of dependencies to the initramfs and
thus was ruled out. It also would introduce GPLv3 into the initramfs.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>