installer image: TPM and encryption optional
Even when enabled in the distro, using a TPM or the well-known
fixed password must be optional: the TPM because the actual
target device might not have a TPM, and the fixed password
because it is insecure.

The installer can be forced to use or not use TPM with
USE_TPM12=true/false. A real installer image should use
USE_TPM12=true when the target hardware has a TPM, to
ensure that problems with using it are detected during
installation instead of silently skipping the use of it.

TODO: use the same values also for the previous boolean env variables.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>